refactor(global): 레포지터리 패키지명 ``repository``로 통일
