+++ src/main/java/team/themoment/hellogsmv3/domain/common/testResult/service/impl/GenerateTestResultCodeServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import team.themoment.hellogsmv3.domain.member.dto.request.GenerateCodeReqDto;
 import team.themoment.hellogsmv3.domain.member.entity.AuthenticationCode;
-import team.themoment.hellogsmv3.domain.member.repo.CodeRepository;
+import team.themoment.hellogsmv3.domain.member.repository.CodeRepository;
 import team.themoment.hellogsmv3.domain.member.service.GenerateCodeService;
 import team.themoment.hellogsmv3.domain.member.service.SendCodeNotificationService;
 import team.themoment.hellogsmv3.global.exception.error.ExpectedException;
