+++ src/test/java/team/themoment/hellogsmv3/domain/common/service/QueryAnnounceTestResultServiceTest.java
@@ -10,7 +10,7 @@ import org.mockito.MockitoAnnotations;
 import org.springframework.http.HttpStatus;
 import team.themoment.hellogsmv3.domain.common.operation.dto.response.AnnounceTestResultResDto;
 import team.themoment.hellogsmv3.domain.common.operation.entity.OperationTestResult;
-import team.themoment.hellogsmv3.domain.common.operation.repo.OperationTestResultRepository;
+import team.themoment.hellogsmv3.domain.common.operation.repository.OperationTestResultRepository;
 import team.themoment.hellogsmv3.domain.common.operation.service.QueryAnnounceTestResultService;
 import team.themoment.hellogsmv3.global.exception.error.ExpectedException;
