+++ src/test/java/team/themoment/hellogsmv3/domain/member/service/impl/GenerateTestCodeServiceImplTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.*;
 import team.themoment.hellogsmv3.domain.member.dto.request.GenerateCodeReqDto;
 import team.themoment.hellogsmv3.domain.member.entity.AuthenticationCode;
-import team.themoment.hellogsmv3.domain.member.repo.CodeRepository;
+import team.themoment.hellogsmv3.domain.member.repository.CodeRepository;
 
 import java.util.Optional;
